Create and test priority queue insertion
I chose to use a priority queue as the data structure to choose the next move.
This is because I'll be using a greedy algorithm and choose the next node to visit
based on what seems the best.

I've also written some tests to make sure the PQ actually works.