Add a function to move one square on a grid and its tests.
moveOnGrid picks the last element from the priority queue, marks the square as visited,
and sorts through its adjacent squares.
We don't want to revisit squares (our score will just keep going down), and we can't
go off the grid. If we find a winner, we'll stop--the first winner we have will be a
best route.
Remaining adjacent squares will get inserted into the priority queue.

We test this with a 3x3 grid for simplicity.

Also:
Add "B" to the effects object since we'll treat it like a blank square.
Make gridSize a global so we can manipulate it from the test file.
(There may be a better way to do this; in an actual unit-testing framework there surely is.)